Turn demo into integration tests
