+++ packages/smoldot-provider/package.json
@@ -28,8 +28,7 @@
   "scripts": {
     "build": "tsc",
     "clean": "rm -rf dist/ && rm -rf node_modules/",
-    "test": "ava",
-    "demo": "DEBUG=smoldot-provider node --experimental-specifier-resolution=node --loader ts-node/esm src/examples/index.ts"
+    "test": "ava"
   },
   "release": [
     "dist/**"
